+++ after/lsp/clangd.lua
@@ -1,0 +1,8 @@
+-- C, C++
+return {
+   cmd = {
+      "clangd",
+      "--header-insertion=never",
+      "--offset-encoding=utf-16",
+   },
+}