+++ after/lsp/eslint.lua
@@ -1,0 +1,16 @@
+-- JavaScript/TypeScript
+return {
+   filetypes = {
+      "javascript",
+      "javascript.jsx",
+      "javascriptreact",
+      "svelte",
+      "typescript",
+      "typescript.tsx",
+      "typescriptreact",
+      "vue",
+   },
+   settings = {
+      workingDirectory = { mode = "auto" },
+   },
+}