+++ after/lsp/julials.lua
@@ -1,0 +1,10 @@
+-- Julia
+return {
+   cmd = {
+      "julia",
+      "--project=@nvim-lspconfig",
+      "-e",
+      "using LanguageServer; using SymbolServer; runserver()",
+   },
+   filetypes = { "julia" },
+}