+++ after/lsp/lua_ls.lua
@@ -1,0 +1,9 @@
+-- Lua
+return {
+   settings = {
+      Lua = {
+         hint = { enable = false },
+         format = { enable = false },
+      },
+   },
+}