+++ after/lsp/rust_analyzer.lua
@@ -1,0 +1,10 @@
+-- Rust
+return {
+   filetypes = { "rust" },
+   root_markers = { "Cargo.toml" },
+   settings = {
+      ["rust-analyzer"] = {
+         cargo = { allFeatures = true },
+      },
+   },
+}