+++ after/lsp/texlab.lua
@@ -1,0 +1,8 @@
+-- LaTeX
+return {
+   settings = {
+      texlab = {
+         latexindent = { modifyLineBreaks = true },
+      },
+   },
+}