+++ after/lsp/vimls.lua
@@ -1,0 +1,2 @@
+-- Vimscript
+return {}